+++ Src/FluentAssertions/Specialized/AssemblyAssertions.cs
@@ -2,6 +2,7 @@
 using System.Collections.Generic;
 using System.Linq;
 using System.Reflection;
+using FluentAssertions.Common;
 using FluentAssertions.Execution;
 using FluentAssertions.Primitives;
 
@@ -31,17 +32,30 @@ namespace FluentAssertions.Reflection
         /// <param name="becauseArgs">
         /// Zero or more objects to format using the placeholders in <paramref name="because" />.
         /// </param>
+        /// <exception cref="ArgumentNullException"><paramref name="assembly"/> is <c>null</c>.</exception>
         public AndConstraint<AssemblyAssertions> NotReference(Assembly assembly, string because = "", params object[] becauseArgs)
         {
-            var subjectName = Subject.GetName().Name;
+            Guard.ThrowIfArgumentIsNull(assembly, nameof(assembly));
+
             var assemblyName = assembly.GetName().Name;
 
-            IEnumerable<string> references = Subject.GetReferencedAssemblies().Select(x => x.Name);
+            bool success = Execute.Assertion
+                .BecauseOf(because, becauseArgs)
+                .ForCondition(Subject is not null)
+                .FailWith("Expected assembly not to reference assembly {0}{reason}, but {context:assembly} is <null>.",
+                    assemblyName);
 
-            Execute.Assertion
-                   .BecauseOf(because, becauseArgs)
-                   .ForCondition(!references.Contains(assemblyName))
-                   .FailWith("Expected assembly {0} not to reference assembly {1}{reason}.", subjectName, assemblyName);
+            if (success)
+            {
+                var subjectName = Subject.GetName().Name;
+
+                IEnumerable<string> references = Subject.GetReferencedAssemblies().Select(x => x.Name);
+
+                Execute.Assertion
+                       .BecauseOf(because, becauseArgs)
+                       .ForCondition(!references.Contains(assemblyName))
+                       .FailWith("Expected assembly {0} not to reference assembly {1}{reason}.", subjectName, assemblyName);
+            }
 
             return new AndConstraint<AssemblyAssertions>(this);
         }
@@ -57,17 +71,29 @@ namespace FluentAssertions.Reflection
         /// <param name="becauseArgs">
         /// Zero or more objects to format using the placeholders in <paramref name="because" />.
         /// </param>
+        /// <exception cref="ArgumentNullException"><paramref name="assembly"/> is <c>null</c>.</exception>
         public AndConstraint<AssemblyAssertions> Reference(Assembly assembly, string because = "", params object[] becauseArgs)
         {
-            var subjectName = Subject.GetName().Name;
+            Guard.ThrowIfArgumentIsNull(assembly, nameof(assembly));
+
             var assemblyName = assembly.GetName().Name;
 
-            IEnumerable<string> references = Subject.GetReferencedAssemblies().Select(x => x.Name);
+            bool success = Execute.Assertion
+                .BecauseOf(because, becauseArgs)
+                .ForCondition(Subject is not null)
+                .FailWith("Expected assembly to reference assembly {0}{reason}, but {context:assembly} is <null>.", assemblyName);
 
-            Execute.Assertion
-                   .BecauseOf(because, becauseArgs)
-                   .ForCondition(references.Contains(assemblyName))
-                   .FailWith("Expected assembly {0} to reference assembly {1}{reason}, but it does not.", subjectName, assemblyName);
+            if (success)
+            {
+                var subjectName = Subject.GetName().Name;
+
+                IEnumerable<string> references = Subject.GetReferencedAssemblies().Select(x => x.Name);
+
+                Execute.Assertion
+                       .BecauseOf(because, becauseArgs)
+                       .ForCondition(references.Contains(assemblyName))
+                       .FailWith("Expected assembly {0} to reference assembly {1}{reason}, but it does not.", subjectName, assemblyName);
+            }
 
             return new AndConstraint<AssemblyAssertions>(this);
         }
@@ -84,15 +110,29 @@ namespace FluentAssertions.Reflection
         /// <param name="becauseArgs">
         /// Zero or more objects to format using the placeholders in <paramref name="because"/>.
         /// </param>
+        /// <exception cref="ArgumentNullException"><paramref name="name"/> is <c>null</c> or empty.</exception>
         public AndWhichConstraint<AssemblyAssertions, Type> DefineType(string @namespace, string name, string because = "", params object[] becauseArgs)
         {
-            Type foundType = Subject.GetTypes().SingleOrDefault(t => t.Namespace == @namespace && t.Name == name);
+            Guard.ThrowIfArgumentIsNullOrEmpty(name, nameof(name));
 
-            Execute.Assertion
-                .ForCondition(foundType is not null)
+            bool success = Execute.Assertion
                 .BecauseOf(because, becauseArgs)
-                .FailWith("Expected assembly {0} to define type {1}.{2}{reason}, but it does not.",
-                    Subject.FullName, @namespace, name);
+                .ForCondition(Subject is not null)
+                .FailWith("Expected assembly to define type {0}.{1}{reason}, but {context:assembly} is <null>.",
+                    @namespace, name);
+
+            Type foundType = null;
+
+            if (success)
+            {
+                foundType = Subject.GetTypes().SingleOrDefault(t => t.Namespace == @namespace && t.Name == name);
+
+                Execute.Assertion
+                    .ForCondition(foundType is not null)
+                    .BecauseOf(because, becauseArgs)
+                    .FailWith("Expected assembly {0} to define type {1}.{2}{reason}, but it does not.",
+                        Subject.FullName, @namespace, name);
+            }
 
             return new AndWhichConstraint<AssemblyAssertions, Type>(this, foundType);
         }