+++ Tests/FluentAssertions.Specs/Specialized/AssemblyAssertionSpecs.cs
@@ -10,6 +10,8 @@ namespace FluentAssertions.Specs.Specialized
 {
     public class AssemblyAssertionSpecs
     {
+        #region NotReference
+
         [Fact]
         public void When_an_assembly_is_not_referenced_and_should_not_reference_is_asserted_it_should_succeed()
         {
@@ -52,6 +54,40 @@ namespace FluentAssertions.Specs.Specialized
             // Assert
             act.Should().Throw<XunitException>();
         }
+
+        [Fact]
+        public void When_subject_is_null_not_reference_should_fail()
+        {
+            // Arrange
+            Assembly assemblyA = null;
+            Assembly assemblyB = FindAssembly.Containing<ClassB>();
+
+            // Act
+            Action act = () => assemblyA.Should().NotReference(assemblyB, "we want to to test the failure {0}", "message");
+
+            // Assert
+            act.Should().Throw<XunitException>()
+                .WithMessage(
+                    "Expected assembly not to reference assembly \"AssemblyB\" *failure message*, but assemblyA is <null>.");
+        }
+
+        [Fact]
+        public void When_an_assembly_is_not_referencing_null_it_should_throw()
+        {
+            // Arrange
+            var assemblyA = FindAssembly.Containing<ClassA>();
+
+            // Act
+            Action act = () => assemblyA.Should().NotReference(null);
+
+            // Assert
+            act.Should().ThrowExactly<ArgumentNullException>()
+                .WithParameterName("assembly");
+        }
+
+        #endregion
+
+        #region Reference
 
         [Fact]
         public void When_an_assembly_is_referenced_and_should_reference_is_asserted_it_should_succeed()
@@ -97,6 +133,40 @@ namespace FluentAssertions.Specs.Specialized
         }
 
         [Fact]
+        public void When_subject_is_null_reference_should_fail()
+        {
+            // Arrange
+            Assembly assemblyA = null;
+            Assembly assemblyB = FindAssembly.Containing<ClassB>();
+
+            // Act
+            Action act = () => assemblyA.Should().Reference(assemblyB, "we want to to test the failure {0}", "message");
+
+            // Assert
+            act.Should().Throw<XunitException>()
+                .WithMessage(
+                    "Expected assembly to reference assembly \"AssemblyB\" *failure message*, but assemblyA is <null>.");
+        }
+
+        [Fact]
+        public void When_an_assembly_is_referencing_null_it_should_throw()
+        {
+            // Arrange
+            var assemblyA = FindAssembly.Containing<ClassA>();
+
+            // Act
+            Action act = () => assemblyA.Should().Reference(null);
+
+            // Assert
+            act.Should().ThrowExactly<ArgumentNullException>()
+                .WithParameterName("assembly");
+        }
+
+        #endregion
+
+        #region DefineType
+
+        [Fact]
         public void When_an_assembly_defines_a_type_and_Should_DefineType_is_asserted_it_should_succeed()
         {
             // Arrange
@@ -129,6 +199,56 @@ namespace FluentAssertions.Specs.Specialized
         }
 
         [Fact]
+        public void When_subject_is_null_define_type_should_fail()
+        {
+            // Arrange
+            Assembly thisAssembly = null;
+
+            // Act
+            Action act = () =>
+                thisAssembly.Should().DefineType(GetType().Namespace, "WellKnownClassWithAttribute",
+                    "we want to to test the failure {0}", "message");
+
+            // Assert
+            act.Should().Throw<XunitException>()
+                .WithMessage(
+                    "Expected assembly to define type *.\"WellKnownClassWithAttribute\" *failure message*" +
+                    ", but thisAssembly is <null>.");
+        }
+
+        [Fact]
+        public void When_an_assembly_defining_a_type_with_a_null_name_it_should_throw()
+        {
+            // Arrange
+            var thisAssembly = GetType().Assembly;
+
+            // Act
+            Action act = () => thisAssembly.Should().DefineType(GetType().Namespace, null);
+
+            // Assert
+            act.Should().ThrowExactly<ArgumentNullException>()
+                .WithParameterName("name");
+        }
+
+        [Fact]
+        public void When_an_assembly_defining_a_type_with_an_empty_name_it_should_throw()
+        {
+            // Arrange
+            var thisAssembly = GetType().Assembly;
+
+            // Act
+            Action act = () => thisAssembly.Should().DefineType(GetType().Namespace, string.Empty);
+
+            // Assert
+            act.Should().ThrowExactly<ArgumentNullException>()
+                .WithParameterName("name");
+        }
+
+        #endregion
+
+        #region BeNull
+
+        [Fact]
         public void When_an_assembly_is_null_and_Should_BeNull_is_asserted_it_should_succeed()
         {
             // Arrange
@@ -141,6 +261,8 @@ namespace FluentAssertions.Specs.Specialized
             // Assert
             act.Should().NotThrow();
         }
+
+        #endregion
     }
 
     [DummyClass("name", true)]